requirements.txt
Flask
PyJWT
cryptography